drm/panel: Add support for In-Cell Synaptics/JDI XXX combo
This adds support for the In-Cell Synaptics TS controller plus
JDI Display XXX (unknown) combo assembly.
It is unclear and nowhere to be found whether this model is
relative to the display itself or to the entire assembly.

This is a 1080x1920@60 display on MIPI DSI, found in some Sony
Xperia X and Sony Xperia X Performance smartphones.